// src/types/generated/AggregationResult.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { AggregatedBuild } from "./AggregatedBuild";
import type { AggregatedChampionRate } from "./AggregatedChampionRate";
import type { AggregatedMatchup } from "./AggregatedMatchup";
import type { AggregationQuality } from "./AggregationQuality";

export type AggregationResult = { rates: Array<AggregatedChampionRate>, matchups: Array<AggregatedMatchup>, builds: Array<AggregatedBuild>, quality: AggregationQuality, };