// src/types/generated/AggregatedMatchup.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type AggregatedMatchup = { champion_id: number, opponent_id: number, position: string, games: number, wins: number, win_rate: number, patch: string, source: string, confidence: string, };