// src/types/generated/AggregatedBuild.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type AggregatedBuild = { champion_id: number, position: string, patch: string, core_items: Array<number>, situational_items: Array<number>, rune_ids: Array<number>, summoner_spells: Array<number>, games: number, win_rate: number, source: string, confidence: string, };